requirements.txt
discord
discord-components
pillow
